restructured har creation, implemented toJson
